backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.4

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.1

# File handling
Pillow==10.2.0

# Server
gunicorn==21.2.0

# Environment variables
python-dotenv==1.0.1

# Utilities
pytz==2024.1
